GMM_Test

+++ DeepNumpy_mlfromscratch/utils/data_operation.py
@@ -1,5 +1,5 @@
 from __future__ import division
-import numpy as np
+from mxnet import np
 import math
 import sys
 
@@ -59,7 +59,7 @@ def calculate_covariance_matrix(X, Y=None):
     n_samples = np.shape(X)[0]
     covariance_matrix = (1 / (n_samples-1)) * (X - X.mean(axis=0)).T.dot(Y - Y.mean(axis=0))
 
-    return np.array(covariance_matrix, dtype=float)
+    return np.array(covariance_matrix, dtype=np.float32)
  
 
 def calculate_correlation_matrix(X, Y=None):
@@ -72,4 +72,4 @@ def calculate_correlation_matrix(X, Y=None):
     std_dev_y = np.expand_dims(calculate_std_dev(Y), 1)
     correlation_matrix = np.divide(covariance, std_dev_X.dot(std_dev_y.T))
 
-    return np.array(correlation_matrix, dtype=float)
+    return np.array(correlation_matrix, dtype=np.float32)
